rym_scraper/requirements.txt
undetected-chromedriver>=3.5.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
